pyproject.toml
[tool.poetry]
name = "strawberry-graphql-mage"
version = "0.0.1-alpha.1"
description = "An automated, modular, backend-agnostic GraphQL schema generator"
license = "MIT"
authors = ["Vojtěch Dohnal <vojdoh@gmail.com>"]
maintainers = []
readme = "README.md"
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]
packages = [
    { include = "strawberry_mage" }
]

[tool.poetry.dependencies]
python = "^3.8"
strawberry-graphql = "^0.84.0"
inflection = "^0.5.1"
starlette = "^0.16.0"
uvicorn = { extras = ["standard"], version = "^0.15.0" }
SQLAlchemy = { extras = ["asyncio"], version = "^1.4.25" }
aiosqlite = "^0.17.0"
frozendict = "^2.0.6"
overrides = "^6.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.16.0"
yappi = "^1.3.3"
nox = "^2021.10.1"
pre-commit = "^2.15.0"
safety = "^1.10.3"
poetry = "^1.1.11"
pydocstyle = {extras = ["toml"], version = "^6.1.1"}
pylint = "^2.11.1"
toml = "^0.10.2"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.poetry.urls]
"Homepage" = "https://github.com/KundaPanda/strawberry-graphql-mage/"


[tool.coverage.run]
include = [
    "strawberry_mage/*"
]
parallel = true

[tool.pyright]
include = ["strawberry_mage"]

[tool.black]
line-length = 120

[tool.pydocstyle]
convention = "pep257"
ignore_decorators = "overrides"

[tool.pylint.master]
ignore = [
    "tests",
]

[tool.pylint.messages_control]
max-line-length = 120
disable = [
    "too-few-public-methods",
    "too-many-return-statements",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
